+++ interfaces/src/main/java/ar/edu/itba/paw/interfaces/services/ExchangeService.java
@@ -10,8 +10,10 @@ public interface ExchangeService {
     void acceptExchange(long acceptCode);
     void rejectExchange(long acceptCode);
     Optional<Exchange> getExchangeById(long exchangeId);
-    long getId(long acceptCode);
-    String exchange(long acceptCode, boolean state);
+
+    long getId(int acceptCode);
+
+    String exchange(int acceptCode, boolean state);
     Exchange initializeExchange(boolean isForExchange, long offererId, long requesterId);
 
 }
